+++ examples/complete/azure/advanced/aks-with-peering/main.tf
@@ -1,0 +1,18 @@
+provider "hopsworksai" {
+}
+
+provider "azurerm" {
+  features {}
+  skip_provider_registration = true
+}
+
+resource "random_string" "suffix" {
+  length  = 8
+  special = false
+  lower   = true
+  upper   = false
+}
+
+locals {
+  suffix_string = random_string.suffix.result
+}