+++ CHANGELOG.md
@@ -5,7 +5,8 @@ NOTES:
 BREAKING CHANGES:
 
 ENHANCEMENTS:
-examples: add advanced azure aks example with vnet peering
+* examples: add advanced azure aks example with vnet peering
+* datasource/aws_instance_profile_policy: Add support for testconnector image
 
 FEATURES: