+++ CHANGELOG.md
@@ -7,7 +7,8 @@ BREAKING CHANGES:
 ENHANCEMENTS:
 * examples: add advanced azure aks example with vnet peering
 * datasource/aws_instance_profile_policy: Add support for testconnector image
-examples: Update terraform helpers modules to version 2.3.0
+* examples: Update terraform helpers modules to version 2.3.0
+* resource/hopsworksai_cluster: remove version suffix before checking
 
 FEATURES: